Move registration API views and serializers to separate app
Refs LINK-1122
